Move test specifications to separate directory
