+++ package.json
@@ -8,7 +8,7 @@
     "build:watch": "node_modules/typescript/bin/tsc --watch",
     "lint": "node_modules/eslint/bin/eslint.js -c ./.eslintrc.js ./src/**/*.ts",
     "lint:fix": "node_modules/eslint/bin/eslint.js -c ./.eslintrc.js --fix ./src/**/*.ts",
-    "test": "node_modules/mocha/bin/mocha -r ts-node/register src/**/*.spec.ts"
+    "test": "node_modules/mocha/bin/mocha -r ts-node/register test/**/*.spec.ts"
   },
   "repository": {
     "type": "git",
