+++ tsconfig.json
@@ -65,6 +65,6 @@
   ],
   "exclude": [
     "node_modules",
-    "**/*.spec.ts"
+    "test"
   ]
 }
